Remove pointless ClientState interface
