+++ server/src/main/java/com/clarkparsia/pellet/server/model/ClientState.java
@@ -8,20 +8,47 @@
 
 package com.clarkparsia.pellet.server.model;
 
+import com.clarkparsia.modularity.IncrementalReasoner;
+import com.clarkparsia.pellet.server.reasoner.LocalSchemaReasoner;
 import com.clarkparsia.pellet.service.reasoner.SchemaReasoner;
+import com.google.common.base.Throwables;
+import org.semanticweb.owlapi.model.OWLOntology;
+import org.semanticweb.owlapi.model.OWLOntologyManager;
 
 /**
  * @author Evren Sirin
  */
-public interface ClientState extends AutoCloseable {
-	SchemaReasoner getReasoner();
+public class ClientState implements AutoCloseable {
+	private final SchemaReasoner reasoner;
+	private final OWLOntology ontology;
+	private final int version;
 
-	/**
-	 * Gets the version of the data used by this reasoner.
-	 *
-	 * @return  an ID of the data version
-	 */
-	int version();
+	public ClientState(final IncrementalReasoner reasoner, final int version) {
+		// create the reasoner with a copy of the incremental reasoner so it won't be affected if the original reasoner is updated
+		this.reasoner = new LocalSchemaReasoner(reasoner.copy());
+		this.ontology = reasoner.getRootOntology();
+		this.version = version;
+	}
 
-	void close();
+	public SchemaReasoner getReasoner() {
+		return reasoner;
+	}
+
+	public int version() {
+		return version;
+	}
+
+	@Override
+	public void close() {
+		try {
+			reasoner.close();
+
+			OWLOntologyManager manager = ontology.getOWLOntologyManager();
+			if (manager != null) {
+				manager.removeOntology(ontology);
+			}
+		} catch (Exception e) {
+			Throwables.propagate(e);
+		}
+	}
 }
